microservicio_consulta/requirements.txt
Django==4.2.24
django-cors-headers>=4.0
requests>=2.31
pandas>=2.0
openpyxl>=3.1
python-dotenv>=1.0
djangorestframework>=3.14
